Serial beginning to work

+++ external/usb/libusb-0.1.12/descriptors.c
@@ -63,7 +63,7 @@ int usb_parse_descriptor(unsigned char *source, char *description, void *dest)
     }
   }
 
-  return sp - source;
+  return (int)sp - (int)source;
 }
 
 /*
